requirements.txt
pyplot
numpy
PIL
psycopg2
python_dotenv
sqlalchemy
pandas